Add PDFgear (#1271)
Hi there. 
I'm Gordon and I just updated the README.md file with the rules provided. I added PDFgear to the Others list in the Reading and Writing Tools catalog. PDFgear is a free AI-integrated PDF software with all the expected functions included, like view, edit, convert, OCR, annotate, etc. The combination of PDF functions and PDFgear Copilot in it makes it super easy to manipulate PDF functions and chat with PDFs. I also added the mark of free and app store after the brief introduction of PDFgear. 
Hope PDFgear could be included in this list after the review. Really appreciate for making this list! Thanks!